digits-android-master/digits/src/main/java/com/digits/sdk/android/LogoImageView.java
/*
 * Copyright (C) 2015 Twitter, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
package com.digits.sdk.android;

import android.content.Context;
import android.graphics.drawable.Drawable;
import android.util.AttributeSet;
import android.view.View;
import android.widget.ImageView;

public class LogoImageView extends ImageView {
    public LogoImageView(Context context) {
        super(context);
        initImageView(context);
    }

    public LogoImageView(Context context, AttributeSet attrs) {
        this(context, attrs, 0);
    }

    public LogoImageView(Context context, AttributeSet attrs, int defStyleAttr) {
        super(context, attrs, defStyleAttr);
        initImageView(context);
    }

    void initImageView(Context context) {
       final Drawable logoDrawable = ThemeUtils.getLogoDrawable(context.getTheme());
        if (logoDrawable != null) {
            setVisibility(View.VISIBLE);
            setImageDrawable(logoDrawable);
        } else {
            setVisibility(View.GONE);
        }
    }

    @Override
    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
        final int width = MeasureSpec.getSize(widthMeasureSpec);
        final int height = width * getDrawable().getIntrinsicHeight() / getDrawable()
                .getIntrinsicWidth();
        setMeasuredDimension(width, height);
    }
}
